requirements.txt
streamlit
grpcio
grpcio-tools
protobuf